Remove custom translations
This was an early hack. We didn't want to wait until Nextcloud had the
updated translations that we had made, so we had the translation files
in here and used those. Nowadays the official translations are better
and more complete, not to mention the fact that this idea was stupid to
begin with.